requirements.txt
ipywidgets==8.1.2
matplotlib==3.8.4
pandas==2.2.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.66.4
kagglehub
flask
tensorflow
numpy
opencv-python-headless
Pillow
requests
google-generativeai
python-dotenv